Beginning of Phase 1 : Dataset section written
I wrote up the data set section based on the bullet points the Jay sent me. I'm still not clear on exactly what you guys did so some of the specifics may be incorrect or require further explanation. Also, I'm not sure how to get the references into BibTex format or if we need them.