Updated Last Paragraph and Other Minor changes
Rewrote the last 2 paragraphs to better explain the process and the
advantage of dropping the indices before the loading of the data. I also
made some other small changes that I found (capitalizing XML and Java,
adding a couple of commas here and there, fixing a misspelled word or
two).